services/tester/requirements.txt
# Alice v2 Testing System Dependencies

# Core framework
asyncio-mqtt==0.16.1
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1

# HTTP/WebSocket communication
httpx==0.27.2
websockets==12.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1  
numpy==1.26.4
scipy==1.13.1

# Data analysis and metrics
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.2

# Testing utilities
pytest==8.3.2
pytest-asyncio==0.24.0

# Chaos engineering
netifaces==0.11.0

# Optional: Advanced audio processing
# whisper-cpp-python==0.2.0  # Uncomment if direct Whisper.cpp integration needed